+++ verticapy/connect/__init__.py
@@ -14,4 +14,24 @@ OR CONDITIONS OF ANY KIND, either express or implied.
 See the  License for the specific  language governing
 permissions and limitations under the License.
 """
-from verticapy.connect.connect import CONNECTION, EXTERNAL_CONNECTION, SPECIAL_SYMBOLS, SESSION_IDENTIFIER, available_connections, change_auto_connection, close_connection, connect, current_connection, current_cursor, delete_connection, get_connection_file, new_connection, read_auto_connect, read_dsn, set_connection, set_external_connection, vertica_connection, verticalab_connection
+from verticapy.connect.connect import (
+    CONNECTION,
+    EXTERNAL_CONNECTION,
+    SPECIAL_SYMBOLS,
+    SESSION_IDENTIFIER,
+    available_connections,
+    change_auto_connection,
+    close_connection,
+    connect,
+    current_connection,
+    current_cursor,
+    delete_connection,
+    get_connection_file,
+    new_connection,
+    read_auto_connect,
+    read_dsn,
+    set_connection,
+    set_external_connection,
+    vertica_connection,
+    verticalab_connection,
+)
