+++ verticapy/datasets/__init__.py
@@ -14,4 +14,24 @@ OR CONDITIONS OF ANY KIND, either express or implied.
 See the  License for the specific  language governing
 permissions and limitations under the License.
 """
-from verticapy.datasets.datasets import *
+from verticapy.datasets.generators import gen_dataset, gen_meshgrid
+from verticapy.datasets.loaders import (
+    load_airline_passengers,
+    load_amazon,
+    load_cities,
+    load_commodities,
+    load_gapminder,
+    load_iris,
+    load_laliga,
+    load_market,
+    load_pop_growth,
+    load_smart_meters,
+    load_titanic,
+    load_winequality,
+    load_world,
+)
+from verticapy.datasets.tests_loaders import (
+    load_dataset_cl,
+    load_dataset_reg,
+    load_dataset_num,
+)
