+++ verticapy/datasets/loaders.py
@@ -20,243 +20,17 @@ permissions and limitations under the License.
 # Modules
 #
 # Standard Python Modules
-import os, datetime
+import os
 
 # VerticaPy Modules
 import verticapy, vertica_python
-from verticapy._version import check_minimum_version
 from verticapy._utils._collect import save_verticapy_logs
 from verticapy.core.vdataframe.vdataframe import vDataFrame
 from verticapy.connect import current_cursor
 from verticapy.sql.create import create_table
 from verticapy.sql.drop import drop
-from verticapy.sql.read import vDataFrameSQL
 from verticapy._utils._sql import _executeSQL
-from verticapy.sql.insert import insert_into
-from verticapy.errors import ParameterError
 from verticapy.sql._utils._format import quote_ident
-
-
-@check_minimum_version
-@save_verticapy_logs
-def gen_dataset(features_ranges: dict, nrows: int = 1000):
-    """
-Generates a dataset using the input parameters.
-
-Parameters
-----------
-features_ranges: dict,
-    Dictionary including the features types and ranges.
-        For str      : The subdictionary must include two keys: 'type' must
-                       be set to 'str' and 'value' must include the feature
-                       categories.
-        For int      : The subdictionary must include two keys: 'type' must
-                       be set to 'int' and 'range' must include two integers
-                       that represent the lower and the upper bound.
-        For float    : The subdictionary must include two keys: 'type' must
-                       be set to 'float' and 'range' must include two floats
-                       that represent the lower and the upper bound.
-        For date     : The subdictionary must include two keys: 'type' must
-                       be set to 'date' and 'range' must include the start
-                       date and the number of days after.
-        For datetime : The subdictionary must include two keys: 'type' must
-                       be set to 'date' and 'range' must include the start
-                       date and the number of days after.
-nrows: int, optional
-    The maximum number of rows in the dataset.
-
-Returns
--------
-vDataFrame
-    Generated dataset.
-    """
-    sql = []
-
-    for param in features_ranges:
-
-        if features_ranges[param]["type"] == str:
-
-            val = features_ranges[param]["values"]
-            if isinstance(val, str):
-                sql += [f"'{val}' AS \"{param}\""]
-            else:
-                n = len(val)
-                val = ", ".join(["'" + str(v) + "'" for v in val])
-                sql += [f'(ARRAY[{val}])[RANDOMINT({n})] AS "{param}"']
-
-        elif features_ranges[param]["type"] == float:
-
-            val = features_ranges[param]["range"]
-            lower, upper = val[0], val[1]
-            sql += [
-                f"""({lower} + RANDOM() 
-                  * ({upper} - {lower}))::FLOAT AS "{param}" """
-            ]
-
-        elif features_ranges[param]["type"] == int:
-
-            val = features_ranges[param]["range"]
-            lower, upper = val[0], val[1]
-            sql += [
-                f"""({lower} + RANDOM() 
-                      * ({upper} - {lower}))::INT AS "{param}" """
-            ]
-
-        elif features_ranges[param]["type"] == datetime.date:
-
-            val = features_ranges[param]["range"]
-            start_date, number_of_days = val[0], val[1]
-            sql += [
-                f"""('{start_date}'::DATE 
-                   + RANDOMINT({number_of_days})) AS "{param}" """
-            ]
-
-        elif features_ranges[param]["type"] == datetime.datetime:
-
-            val = features_ranges[param]["range"]
-            start_date, number_of_days = val[0], val[1]
-            sql += [
-                f"""('{start_date}'::TIMESTAMP 
-                   + {number_of_days} * RANDOM()) AS "{param}" """
-            ]
-
-        elif features_ranges[param]["type"] == bool:
-
-            sql += [f'RANDOMINT(2)::BOOL AS "{param}"']
-
-        else:
-
-            ptype = features_ranges[param]["type"]
-            raise ParameterError(f"Parameter {param}: Type {ptype} is not supported.")
-
-    query = f"""
-        (SELECT {', '.join(sql)} FROM 
-            (SELECT tm FROM 
-                (SELECT '03-11-1993'::TIMESTAMP + INTERVAL '1 second' AS t 
-                 UNION ALL 
-                 SELECT '03-11-1993'::TIMESTAMP + INTERVAL '{nrows} seconds' AS t) x 
-                TIMESERIES tm AS '1 second' OVER(ORDER BY t)) y) z"""
-
-    return vDataFrameSQL(query)
-
-
-@save_verticapy_logs
-def gen_meshgrid(features_ranges: dict):
-    """
-Generates a dataset using regular steps.
-
-Parameters
-----------
-features_ranges: dict,
-    Dictionary including the features types and ranges.
-        For str      : The subdictionary must include two keys: 'type' must
-                       be set to 'str' and 'value' must include the feature
-                       categories.
-        For int      : The subdictionary must include two keys: 'type' must
-                       be set to 'int' and 'range' must include two integers
-                       that represent the lower and the upper bound.
-        For float    : The subdictionary must include two keys: 'type' must
-                       be set to 'float' and 'range' must include two floats
-                       that represent the lower and the upper bound.
-        For date     : The subdictionary must include two keys: 'type' must
-                       be set to 'date' and 'range' must include the start
-                       date and the number of days after.
-        For datetime : The subdictionary must include two keys: 'type' must
-                       be set to 'date' and 'range' must include the start
-                       date and the number of days after.
-        Numerical and date-like features must have an extra key in the 
-        dictionary named 'nbins' corresponding to the number of bins used to 
-        compute the different categories.
-
-Returns
--------
-vDataFrame
-    generated dataset.
-    """
-    sql = []
-
-    for idx, param in enumerate(features_ranges):
-
-        nbins = 100
-        if "nbins" in features_ranges[param]:
-            nbins = features_ranges[param]["nbins"]
-        ts_table = f"""
-            (SELECT 
-                DAY(tm - '03-11-1993'::TIMESTAMP) AS tm 
-             FROM 
-                (SELECT '03-11-1993'::TIMESTAMP AS t 
-                 UNION ALL 
-                 SELECT '03-11-1993'::TIMESTAMP 
-                        + INTERVAL '{nbins} days' AS t) x 
-            TIMESERIES tm AS '1 day' OVER(ORDER BY t)) y"""
-
-        if features_ranges[param]["type"] == str:
-            val = features_ranges[param]["values"]
-            if isinstance(val, str):
-                val = [val]
-            val = " UNION ALL ".join([f"""(SELECT '{v}' AS "{param}")""" for v in val])
-            sql += [f"({val}) x{idx}"]
-
-        elif features_ranges[param]["type"] == float:
-            val = features_ranges[param]["range"]
-            lower, upper = val[0], val[1]
-            h = (upper - lower) / nbins
-            sql += [
-                f"""
-                (SELECT 
-                    ({lower} + {h} * tm)::FLOAT AS "{param}" 
-                 FROM {ts_table}) x{idx}"""
-            ]
-
-        elif features_ranges[param]["type"] == int:
-            val = features_ranges[param]["range"]
-            lower, upper = val[0], val[1]
-            h = (upper - lower) / nbins
-            sql += [
-                f"""
-                (SELECT 
-                    ({lower} + {h} * tm)::INT AS "{param}" 
-                 FROM {ts_table}) x{idx}"""
-            ]
-
-        elif features_ranges[param]["type"] == datetime.date:
-            val = features_ranges[param]["range"]
-            start_date, number_of_days = val[0], val[1]
-            h = number_of_days / nbins
-            sql += [
-                f"""
-                (SELECT 
-                    ('{start_date}'::DATE + {h} * tm)::DATE AS "{param}" 
-                 FROM {ts_table}) x{idx}"""
-            ]
-
-        elif features_ranges[param]["type"] == datetime.datetime:
-            val = features_ranges[param]["range"]
-            start_date, number_of_days = val[0], val[1]
-            h = number_of_days / nbins
-            sql += [
-                f"""
-                    (SELECT 
-                        ('{start_date}'::DATE + {h} * tm)::TIMESTAMP 
-                            AS "{param}" 
-                     FROM {ts_table}) x{idx}"""
-            ]
-
-        elif features_ranges[param]["type"] == bool:
-            sql += [
-                f"""
-                ((SELECT False AS "{param}") 
-                 UNION ALL
-                (SELECT True AS "{param}")) x{idx}"""
-            ]
-
-        else:
-            ptype = features_ranges[param]["type"]
-            raise ParameterError(f"Parameter {param}: Type {ptype} is not supported.")
-
-    query = f"(SELECT * FROM {' CROSS JOIN '.join(sql)}) x"
-
-    return vDataFrameSQL(query)
 
 
 def load_dataset(
@@ -817,112 +591,3 @@ vDataFrame
         ],
         dataset_name="world",
     )
-
-
-#
-# Datasets used in the tests
-#
-
-
-def load_dataset_cl(table_name: str = "dataset_cl", schema: str = "public"):
-    # Classification Dataset
-
-    data = [
-        [1, "Bus", "Male", 0, "Cheap", "Low"],
-        [2, "Bus", "Male", 1, "Cheap", "Med"],
-        [3, "Train", "Female", 1, "Cheap", "Med"],
-        [4, "Bus", "Female", 0, "Cheap", "Low"],
-        [5, "Bus", "Male", 1, "Cheap", "Med"],
-        [6, "Train", "Male", 0, "Standard", "Med"],
-        [7, "Train", "Female", 1, "Standard", "Med"],
-        [8, "Car", "Female", 1, "Expensive", "Hig"],
-        [9, "Car", "Male", 2, "Expensive", "Med"],
-        [10, "Car", "Female", 2, "Expensive", "Hig"],
-    ]
-    input_relation = f"{quote_ident(schema)}.{quote_ident(table_name)}"
-
-    drop(name=input_relation, method="table")
-    create_table(
-        table_name=table_name,
-        schema=schema,
-        dtype={
-            "Id": "INT",
-            "transportation": "VARCHAR",
-            "gender": "VARCHAR",
-            "owned cars": "INT",
-            "cost": "VARCHAR",
-            "income": "CHAR(4)",
-        },
-    )
-    insert_into(table_name=table_name, schema=schema, data=data, copy=False)
-
-    return vDataFrame(input_relation=input_relation)
-
-
-def load_dataset_reg(table_name: str = "dataset_reg", schema: str = "public"):
-    # Regression Dataset
-
-    data = [
-        [1, 0, "Male", 0, "Cheap", "Low"],
-        [2, 0, "Male", 1, "Cheap", "Med"],
-        [3, 1, "Female", 1, "Cheap", "Med"],
-        [4, 0, "Female", 0, "Cheap", "Low"],
-        [5, 0, "Male", 1, "Cheap", "Med"],
-        [6, 1, "Male", 0, "Standard", "Med"],
-        [7, 1, "Female", 1, "Standard", "Med"],
-        [8, 2, "Female", 1, "Expensive", "Hig"],
-        [9, 2, "Male", 2, "Expensive", "Med"],
-        [10, 2, "Female", 2, "Expensive", "Hig"],
-    ]
-    input_relation = f"{quote_ident(schema)}.{quote_ident(table_name)}"
-
-    drop(name=input_relation, method="table")
-    create_table(
-        table_name=table_name,
-        schema=schema,
-        dtype={
-            "Id": "INT",
-            "transportation": "INT",
-            "gender": "VARCHAR",
-            "owned cars": "INT",
-            "cost": "VARCHAR",
-            "income": "CHAR(4)",
-        },
-    )
-    insert_into(table_name=table_name, schema=schema, data=data, copy=False)
-
-    return vDataFrame(input_relation=input_relation)
-
-
-def load_dataset_num(table_name: str = "dataset_num", schema: str = "public"):
-    # Numerical Dataset
-
-    data = [
-        [1, 7.2, 3.6, 6.1, 2.5],
-        [2, 7.7, 2.8, 6.7, 2.0],
-        [3, 7.7, 3.0, 6.1, 2.3],
-        [4, 7.9, 3.8, 6.4, 2.0],
-        [5, 4.4, 2.9, 1.4, 0.2],
-        [6, 4.6, 3.6, 1.0, 0.2],
-        [7, 4.7, 3.2, 1.6, 0.2],
-        [8, 6.5, 2.8, 4.6, 1.5],
-        [9, 6.8, 2.8, 4.8, 1.4],
-        [10, 7.0, 3.2, 4.7, 1.4],
-    ]
-    input_relation = f"{quote_ident(schema)}.{quote_ident(table_name)}"
-
-    drop(name=input_relation, method="table")
-    create_table(
-        table_name=table_name,
-        schema=schema,
-        dtype={
-            "Id": "INT",
-            "col1": "FLOAT",
-            "col2": "FLOAT",
-            "col3": "FLOAT",
-            "col4": "FLOAT",
-        },
-    )
-    insert_into(table_name=table_name, schema=schema, data=data, copy=False)
-
-    return vDataFrame(input_relation=input_relation)
