+++ verticapy/errors/__init__.py
@@ -15,11 +15,6 @@ See the  License for the specific  language governing
 permissions and limitations under the License.
 """
 
-#
-#
-# Errors & Exceptions
-#
-
 
 class ConnectionError(Exception):
     ...
