+++ verticapy/tests/connect/test_connect.py
@@ -14,8 +14,7 @@ OR CONDITIONS OF ANY KIND, either express or implied.
 See the  License for the specific  language governing
 permissions and limitations under the License.
 """
-
-# VerticaPy
+import os
 import verticapy as vp
 from verticapy.connect import *
