+++ verticapy/tests/utilities/test_utilities.py
@@ -16,7 +16,7 @@ permissions and limitations under the License.
 """
 
 # Pytest
-import pytest
+import pytest, warnings
 
 # Other Modules
 import pandas as pd
